requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.10
typing-extensions>=4.8.0
aiohttp>=3.9.0
pydantic>=2.0.0
beautifulsoup4>=4.12.0  # For web scraping
requests>=2.31.0
python-dotenv>=1.0.0  # For environment variables
google-genai
pymongo
langgraph-checkpoint-mongodb